Fix BuildResultPopupAssetStrategy with tests
